requirements.txt
# Data
yfinance>=0.2.31
pandas>=2.1.0
numpy>=1.24.0

# ML / Deep Learning
scikit-learn>=1.3.0
tensorflow>=2.14.0
keras>=3.0.0

# Visualization
plotly>=5.18.0
matplotlib>=3.8.0

# Dashboard
streamlit>=1.29.0

# Config
pyyaml>=6.0

# Testing
pytest>=7.4.0

# Utilities
scipy>=1.11.0
ta>=0.11.0
